Fix + expand CS2-SDK-Reference: source2sdk (was Dota), schema (3→18 modules), tooling, docs

+++ source2sdk/sdk/include/source2sdk/modellib/InputLayoutVariation_t.hpp
@@ -14,9 +14,9 @@ namespace source2sdk
     namespace modellib
     {
         // Enumerator count: 4
-        // Alignment: 4
-        // Size: 0x4
-        enum class InputLayoutVariation_t : std::uint32_t
+        // Alignment: 1
+        // Size: 0x1
+        enum class InputLayoutVariation_t : std::uint8_t
         {
             INPUT_LAYOUT_VARIATION_DEFAULT = 0x0,
             INPUT_LAYOUT_VARIATION_STREAM1_INSTANCEID = 0x1,
